UserController Refactoring
